Undoing exthost side of the webview serializer split
